do some refactorings
